updated menubar with more options
They still need a little fine tuning, but added auto on/off for text, and allowing users to choose faster or slower pace for lines to be displayed.  currentscript is being incremented with any click in gamestate 5, so I'll need to figure out how to set that to only happen outside of the existing buttons.