chapter2 page 8 parallax images and code added
last page of chapter 2 has 4 layers of it's parallax instead of just 3 compared to the page before it.  code may be ready to be refactored a bit, and debugging definitely needs to take place, but chapter 2 art, dialog, and functionality have been implemented. I may still reread the dialog and make some changes though.  probably going to update the buttons to include chapter titles instead of just chapter numbers.