bug fixes, tin can.
one of the background images in chapter 3 was throwing a weird warning to the console,
found the offending image and corrected the problem (reexported it using different color profile).
also added a little tin can image to the lower left of the dialog box.
also there was a frame of the rat animation that never got exported, so I included that as well.